Allow compiling for web, although doesn't properly function

+++ src/sdl/KeybindingsMenu.cpp
@@ -50,7 +50,7 @@ void KeybindingsMenu::init() {
 
     const uint64_t totalMemorySize = Clay_MinMemorySize();
     const Clay_Arena clayMemory = {
-        .capacity = totalMemorySize,
+        .capacity = static_cast<size_t>(totalMemorySize),
         .memory = static_cast<char*>(SDL_malloc(totalMemorySize))
     };
